Fetch sst for sensor surveys from time series
