+++ packages/api/src/sensors/sensors.module.ts
@@ -1,6 +1,5 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { DailyData } from '../reefs/daily-data.entity';
 import { Reef } from '../reefs/reefs.entity';
 import { Sources } from '../reefs/sources.entity';
 import { Survey } from '../surveys/surveys.entity';
@@ -11,8 +10,6 @@ import { SensorsService } from './sensors.service';
 @Module({
   controllers: [SensorsController],
   providers: [SensorsService],
-  imports: [
-    TypeOrmModule.forFeature([DailyData, Reef, Sources, Survey, TimeSeries]),
-  ],
+  imports: [TypeOrmModule.forFeature([Reef, Sources, Survey, TimeSeries])],
 })
 export class SensorsModule {}
