+++ packages/api/src/sensors/sensors.service.ts
@@ -1,7 +1,7 @@
 import { BadRequestException, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import Bluebird from 'bluebird';
-import { groupBy, keyBy, mapValues } from 'lodash';
+import { keyBy } from 'lodash';
 import { GeoJSON, Point } from 'geojson';
 import { IsNull, Not, Repository } from 'typeorm';
 import { Reef, SensorType } from '../reefs/reefs.entity';
@@ -13,21 +13,16 @@ import { getSpotterData, getLatestData } from '../utils/sofar';
 import { createPoint } from '../utils/coordinates';
 import { SpotterData } from '../utils/sofar.types';
 import {
-  TimeSeriesData,
   getDataQuery,
   groupByMetricAndSource,
+  getClosestTimeSeriesData,
 } from '../utils/time-series.utils';
-import { DailyData } from '../reefs/daily-data.entity';
 import { Sources } from '../reefs/sources.entity';
-import { SensorDataDto } from './dto/sensor-data.dto';
 import { SourceType } from '../reefs/schemas/source-type.enum';
 
 @Injectable()
 export class SensorsService {
   constructor(
-    @InjectRepository(DailyData)
-    private dailyDataRepository: Repository<DailyData>,
-
     @InjectRepository(Reef)
     private reefRepository: Repository<Reef>,
 
@@ -131,16 +126,17 @@ export class SensorsService {
 
     return Promise.all(
       surveyDetails.map(async (survey) => {
-        const reefTimeSeries = await this.getClosestTimeSeriesData(
+        const reefSensorData = await getClosestTimeSeriesData(
           survey.diveDate,
           survey.reefId,
-          [Metric.BOTTOM_TEMPERATURE, Metric.TOP_TEMPERATURE],
-          [SourceType.SPOTTER],
-        );
-
-        const dailyData = await this.getClosestDailyData(
-          survey.diveDate,
-          survey.reefId,
+          [
+            Metric.BOTTOM_TEMPERATURE,
+            Metric.TOP_TEMPERATURE,
+            Metric.SATELLITE_TEMPERATURE,
+          ],
+          [SourceType.SPOTTER, SourceType.NOAA],
+          this.timeSeriesRepository,
+          this.sourcesRepository,
         );
 
         const surveyMedia = await Promise.all(
@@ -149,11 +145,13 @@ export class SensorsService {
               return media;
             }
 
-            const poiTimeSeries = await this.getClosestTimeSeriesData(
+            const poiTimeSeries = await getClosestTimeSeriesData(
               survey.diveDate,
               survey.reefId,
               [Metric.BOTTOM_TEMPERATURE, Metric.TOP_TEMPERATURE],
               [SourceType.HOBO],
+              this.timeSeriesRepository,
+              this.sourcesRepository,
               media.poiId,
             );
 
@@ -167,113 +165,9 @@ export class SensorsService {
         return {
           ...survey,
           surveyMedia,
-          sensorData: {
-            ...reefTimeSeries,
-            ...dailyData,
-          },
+          sensorData: reefSensorData,
         };
       }),
     );
-  }
-
-  private async getClosestDailyData(
-    diveDate: Date,
-    reefId: number,
-  ): Promise<SensorDataDto | {}> {
-    // We will use this many times in our query, so we declare it as constant
-    const diff = `(daily_data.date::timestamp - '${diveDate.toISOString()}'::timestamp)`;
-
-    // We order (ascending) the data by the date difference between the date column and diveData
-    // and we grab the first one, which will be the closest one
-    const dailyData = await this.dailyDataRepository
-      .createQueryBuilder('daily_data')
-      .where('daily_data.reef_id = :reefId', { reefId })
-      .andWhere(`${diff} < INTERVAL '1 d'`)
-      .andWhere(`${diff} > INTERVAL '-1 d'`)
-      .orderBy(
-        `(CASE WHEN ${diff} < INTERVAL '0' THEN (-${diff}) ELSE ${diff} END)`,
-        'ASC',
-      )
-      .getOne();
-
-    if (!dailyData) {
-      return {};
-    }
-
-    // Create a SensorData object that contains the data point
-    return {
-      [SourceType.NOAA]: {
-        [Metric.SATELLITE_TEMPERATURE]: {
-          value: dailyData.satelliteTemperature,
-          timestamp: dailyData.date,
-        },
-      },
-    };
-  }
-
-  private async getClosestTimeSeriesData(
-    diveDate: Date,
-    reefId: number,
-    metrics: Metric[],
-    sourceTypes: SourceType[],
-    poiId?: number,
-  ) {
-    const poiCondition = poiId
-      ? `source.poi_id = ${poiId}`
-      : 'source.poi_id IS NULL';
-    // We will use this many times in our query, so we declare it as constant
-    const diff = `(time_series.timestamp::timestamp - '${diveDate.toISOString()}'::timestamp)`;
-
-    // First get all sources needed to avoid inner join later
-    const sources = await this.sourcesRepository
-      .createQueryBuilder('source')
-      .where('source.type IN (:...sourceTypes)', { sourceTypes })
-      .andWhere('source.reef_id = :reefId', { reefId })
-      .andWhere(poiCondition)
-      .getMany();
-
-    if (!sources.length) {
-      return {};
-    }
-
-    // Create map from source_id to source entity
-    const sourceMap = keyBy(sources, (source) => source.id);
-
-    // Grab all data at an interval of +/- 24 hours around the diveDate
-    // Order (descending) those data by the absolute time distance between the data and the survey diveDate
-    // This way the closest data point for each metric for each source type will be the last row
-    const timeSeriesData: TimeSeriesData[] = await this.timeSeriesRepository
-      .createQueryBuilder('time_series')
-      .select('time_series.timestamp', 'timestamp')
-      .addSelect('time_series.value', 'value')
-      .addSelect('time_series.metric', 'metric')
-      .addSelect('time_series.source_id', 'source')
-      .where(`${diff} < INTERVAL '1 d'`)
-      .andWhere(`${diff} > INTERVAL '-1 d'`)
-      .andWhere('time_series.metric IN (:...metrics)', { metrics })
-      .andWhere('time_series.source_id IN (:...sourceIds)', {
-        sourceIds: Object.keys(sourceMap),
-      })
-      .orderBy(
-        `time_series.source_id, metric, (CASE WHEN ${diff} < INTERVAL '0' THEN (-${diff}) ELSE ${diff} END)`,
-        'DESC',
-      )
-      .getRawMany();
-
-    // Group the data by source id
-    const groupedData = groupBy(timeSeriesData, (o) => o.source);
-
-    return Object.keys(groupedData).reduce<SensorDataDto>((data, key) => {
-      return {
-        ...data,
-        // Replace source id by source using the mapped source object
-        // Keep only timestamps and value from the resulting objects
-        [sourceMap[key].type]: mapValues(
-          // Use key by to group the data by metric and keep only the last entry, i.e. the closest one
-          keyBy(groupedData[key], (grouped) => grouped.metric),
-          (v) => ({ timestamp: v.timestamp, value: v.value }),
-        ),
-      };
-    }, {});
   }
 }
